# shared/contracts/auth.openapi.yaml
openapi: 3.0.3
info:
  title: SCM Auth Service API
  version: 0.1.0
  description: Authentication and session management contract.
servers:
  - url: /api/auth
paths:
  /v1/login:
    post:
      summary: Login and issue access token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
  /v1/tokens/verify:
    post:
      summary: Verify access token (introspection style)
      operationId: verifyToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTokenRequest"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyTokenResponse"
components:
  schemas:
    LoginRequest:
      type: object
      required: [loginId, password]
      properties:
        loginId:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresIn, expiresAt, memberId, roles]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
          minimum: 1
        expiresAt:
          type: string
          format: date-time
        memberId:
          type: string
        roles:
          type: array
          items:
            type: string
    VerifyTokenRequest:
      type: object
      required: [accessToken]
      properties:
        accessToken:
          type: string
    VerifyTokenResponse:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
        subject:
          type: string
        roles:
          type: array
          items:
            type: string
        issuedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time